% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer_adaptive_normalizer.R
\name{AdaptiveNormalizer}
\alias{AdaptiveNormalizer}
\title{Adaptive Normalizer}
\description{
R6 class with our adaptation of the method proposed by Ogasawara et al. for one-pass online
time-series adaptive normalization. This method is designed to train neural networks and returns
the normalized train and test feature sets. In this proposal, the outlier elimination phase has
been omitted.
}
\examples{

normalizer <- AdaptiveNormalizer$new(3)

normalizer$normalize(10)
normalizer$normalize(15)
normalizer$normalize(20)
normalizer$normalize(10)
normalizer$normalize(30)
normalizer$normalize(15)

normalizer$denormalize(-0.4858841)

}
\references{
E. Ogasawara, L. C. Martinez, D. De Oliveira, G. Zimbrão, G. L. Pappa,M. Mattoso,
Adaptive Normalization: A novel data normalization ap-proach for non-stationary time series,
in: Proceedings of the International Joint Conference on Neural Networks, 2010, pp. 1-8
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AdaptiveNormalizer$new()}}
\item \href{#method-normalize}{\code{AdaptiveNormalizer$normalize()}}
\item \href{#method-denormalize}{\code{AdaptiveNormalizer$denormalize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new AdaptiveNormalizer2 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveNormalizer$new(wl = 10, maxmin = T, l = 3, returnPoint = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wl}}{Length of the window with historical data values.}

\item{\code{maxmin}}{If \code{TRUE} standardization and normalization are applied, else only
standardization is applied.}

\item{\code{l}}{Times IQR to anomaly removing. It must be a number greater than 0. By default 3,
but other common values could be 1.5 and 6.}

\item{\code{returnPoint}}{If \code{FALSE} normalized then the current window is returned, else, the
normalized current data point is returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new AdaptiveNormalizer object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-normalize"></a>}}
\if{latex}{\out{\hypertarget{method-normalize}{}}}
\subsection{Method \code{normalize()}}{
Normalizes the current data value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveNormalizer$normalize(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Current data value to be normalized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If \code{returnPoint = FALSE} then normalized current train and test windows are
returned, else, normalized current data point is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-denormalize"></a>}}
\if{latex}{\out{\hypertarget{method-denormalize}{}}}
\subsection{Method \code{denormalize()}}{
Denormalizes the current data value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveNormalizer$denormalize(y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{Current data value to be denormalized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Denormalized current data point.
}
}
}
