% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer_adaptive_normalizer2.R
\name{AdaptiveNormalizer2}
\alias{AdaptiveNormalizer2}
\title{One-pass adaptive min-max normalizalizer}
\description{
R6 class that implements our adaptation to one-pass online processing of the adaptive
normalization method proposed by Gupta and Hewett. A sliding window of past \code{w} data values
are maintained and the percentage change in mean values between the current and previous window
is used to determine if the new or old min-max values should be used to normalize the current
window. If the percentage change value exceeds the threshold, then old min-max values are
replaced by the new ones.
}
\examples{

normalizer <- AdaptiveNormalizer2$new(wl = 3)

normalizer$normalize(10)
normalizer$normalize(15)
normalizer$normalize(20)
normalizer$normalize(10)
normalizer$normalize(30)
normalizer$normalize(15)

normalizer$denormalize(0.25)

}
\references{
V. Gupta, R. Hewett, Adaptive Normalization in Streaming Data, in:Proceedings of
the 2019 3rd International Conference on Big Data Re-search, ACM, New York, NY, USA, 2019,
pp. 12-17.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AdaptiveNormalizer2$new()}}
\item \href{#method-normalize}{\code{AdaptiveNormalizer2$normalize()}}
\item \href{#method-denormalize}{\code{AdaptiveNormalizer2$denormalize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new AdaptiveNormalizalizer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveNormalizer2$new(wl, threshold = 0.25, returnPoint = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wl}}{Length of the window with historical data values.}

\item{\code{threshold}}{Dissimilarity factor between the previous and current windows mean to
calculate new min and max values.}

\item{\code{returnPoint}}{If \code{FALSE} then normalized current window is returned, else,
the normalized current data point is returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `AdaptiveNormalizer2` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-normalize"></a>}}
\if{latex}{\out{\hypertarget{method-normalize}{}}}
\subsection{Method \code{normalize()}}{
Normalizes the current data value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveNormalizer2$normalize(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Current data value to be normalized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If \code{returnPoint = FALSE} then normalized current window is returned, else,
normalized current data point is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-denormalize"></a>}}
\if{latex}{\out{\hypertarget{method-denormalize}{}}}
\subsection{Method \code{denormalize()}}{
Denormalizes the current data value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveNormalizer2$denormalize(y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{Current data value to be denormalized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Denormalized current data point.
}
}
}
