% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ambient_temperature_system_failure}
\alias{ambient_temperature_system_failure}
\title{ambient_temperature_system_failure.}
\format{A data frame with three variables: \code{timestamp}, \code{value},
  \code{is.real.anomaly}.

For further details, see \url{https://github.com/numenta/NAB/blob/master/data/README.md}}
\usage{
ambient_temperature_system_failure
}
\description{
The ambient temperature in an office setting.
}
\keyword{datasets}
