% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{art_daily_jumpsup}
\alias{art_daily_jumpsup}
\title{art_daily_jumpsup}
\format{
A data frame with three variables: \code{timestamp}, \code{value},
  \code{is.real.anomaly}.

For further details, see \url{https://github.com/numenta/NAB/blob/master/data/README.md}
}
\usage{
art_daily_jumpsup
}
\description{
Artificially-generated data with varying types of anomalies
}
\keyword{datasets}
