% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextCrosser.CAD.R
\name{ContextCrosser.CAD}
\alias{ContextCrosser.CAD}
\title{Context Crosser for Contextual Anomaly Detector (ContextualAD)}
\usage{
ContextCrosser.CAD(left.or.right, fact.list, new.ctxt.flag = FALSE,
  potential.new.contexts = list(), context.operator)
}
\arguments{
\item{left.or.right}{Situation from which the method have been
called. It says if the new facts have been seen or if they are
completely new.}

\item{fact.list}{New fact list.}

\item{new.ctxt.flag}{Flag that indicates if a new context is required.}

\item{potential.new.contexts}{Potential new contexts that might be added.}
}
\value{
If \code{left.or.right}, number of new contex added tho the
    \code{context.operator}. Otherwise, the output from
    \code{UpdateContextsAndGetActive.CAD}.
}
\description{
\code{ContextCrosser.CAD} is an auxiliar method that makes the necessary
changes in the \code{context.operator} environment, including in
the \code{fact.list} and the corresponding dictionary of facts the
new facts.
}
\details{
\code{left.or.right} must be 0 or 1 to indicate left or
    right respectively.
}
\references{
Smirnov, M. (2018). CAD: Contextual Anomaly
    Detector. https://github.com/smirmik/CAD
}
