% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_pewma.R
\name{CpPewma}
\alias{CpPewma}
\title{Classic Processing Probabilistic-EWMA (PEWMA)}
\usage{
CpPewma(data, n.train = 5, alpha0 = 0.2, beta = 0, l = 3)
}
\arguments{
\item{data}{Numeric vector that includes train and test data to analyze.}

\item{n.train}{Number of points of the data set that correspond to the training set.}

\item{alpha0}{initial alpha value}

\item{beta}{beta value}

\item{l}{Sigma multiplier to calculate the control limits.}
}
\value{
Data set conformed by the following columns:

  \item{is.anomaly}{1 if the value is anomalous 0 otherwise.}
  \item{ucl}{Upper control limit.}
  \item{lcl}{Lower control limit.}
}
\description{
\code{CpPewma} calculates the anomalies of a data set using classical
processing based on the PEWMA algorithm.
}
\details{
\code{data} must be numerical vector without NA values. BLA BLA BLA
Finally, \code{l} is the parameter that determines the control limits.
By default, 3 is used.
}
\examples{
## Generate data
set.seed(100)
n <- 500
x <- sample(1:100, n, replace = TRUE)
x[70:90] <- sample(110:115, 21, replace = TRUE)
x[25] <- 200
x[320] <- 170
df <- data.frame(timestamp=1:n,value=x)

## Calculate anomalies
result <- CpPewma(
  data = df$value,
  n.train = 5,
  alpha0 = 0.8,
  beta = 0.3,
  l = 3
)

## Plot results
res <- cbind(df, result)
res <- res[1:500,]
y.limits <- c(-150,250)
plot(x = res$timestamp, y = res$value, type = "l", ylim = y.limits, xlab = "timestamp", ylab = "value", main = "Pewma ANOMALY DETECTOR")
points(x = res[res$is.anomaly == 1, "timestamp"], y = res[res$is.anomaly == 1, "value"], pch=4, col="red", lwd = 2)
par(new=TRUE)
plot(x = res$timestamp, y = res$ucl, type="l", col="green", xaxt="n", ylim = y.limits, xlab = "", ylab = "")
par(new=TRUE)
plot(x = res$timestamp, y = res$lcl, type="l", col="green", xaxt="n", ylim = y.limits, xlab = "", ylab = "")

}
\references{
BLA BLA BLA
}
