% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer_dynamic_normalizer.R
\name{DynamicNormalizer}
\alias{DynamicNormalizer}
\title{Dinamyc normalizer}
\description{
R6 class that implements the unsupervised dynamic z-score standardization proposed by Bollegala.
This method allows normalizing the data sets in an online manner, by z-score standardization,
dynamically updating the mean and the variance whenever a new data arrives.
}
\examples{

normalizer <- DynamicNormalizer$new()

normalizer$normalize(10)
normalizer$normalize(15)
normalizer$normalize(20)
normalizer$normalize(10)
normalizer$normalize(30)
normalizer$normalize(15)

normalizer$denormalize(-0.2214036)

}
\references{
D. Bollegala, Dynamic feature scaling for online learning of binary classifiers,
Knowledge-Based Syst., vol. 129, pp. 97–105, 2017.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DynamicNormalizer$new()}}
\item \href{#method-normalize}{\code{DynamicNormalizer$normalize()}}
\item \href{#method-denormalize}{\code{DynamicNormalizer$denormalize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new DynamicNormalizer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicNormalizer$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new DynamicNormalizer object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-normalize"></a>}}
\if{latex}{\out{\hypertarget{method-normalize}{}}}
\subsection{Method \code{normalize()}}{
Normalizes the current data value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicNormalizer$normalize(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Current data value to be normalized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If \code{return_point = FALSE} then normalized current window is returned, else,
normalized current data point is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-denormalize"></a>}}
\if{latex}{\out{\hypertarget{method-denormalize}{}}}
\subsection{Method \code{denormalize()}}{
Denormalizes the current data value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicNormalizer$denormalize(y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{Current data value to be denormalized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Denormalized current data point.
}
}
}
