% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_dynamic_sigma_scorer.R
\name{DynamicSigmaScorer}
\alias{DynamicSigmaScorer}
\title{Dynamic Sigma Scorer}
\description{
R6 class that allows computing the anomaly score online based on the prediction errors and
3-sigma control limits. The mean and standard deviations for 3-sima control limits are
dynamically computed whenever new data arrives.
}
\examples{

scorer <- DynamicSigmaScorer$new()

scorer$computeScore(0.5)
scorer$computeScore(0.6)
scorer$computeScore(0.5)
scorer$computeScore(7)
scorer$computeScore(0.2)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DynamicSigmaScorer$new()}}
\item \href{#method-computeScore}{\code{DynamicSigmaScorer$computeScore()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new DynamicSigmaScorer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicSigmaScorer$new(l = 3, computeLogLikelihood = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{l}}{Number of deviations from the mean.}

\item{\code{computeLogLikelihood}}{If \code{TRUE} compute a log scale representation of the
likelihood value. Since thelikelihood computations return low probabilities that often go
into four 9's or five 9's, a log value is more useful for visualization, thresholding, etc.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new DynamicSigmaScorer object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-computeScore"></a>}}
\if{latex}{\out{\hypertarget{method-computeScore}{}}}
\subsection{Method \code{computeScore()}}{
Calculates the anomaly score from the prediction error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicSigmaScorer$computeScore(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Current error value to be scored.}

\item{\code{...}}{Any other parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Anomaly score, a value between [0,1]. The closer the number is
to 1, the higher the chance it is an anomaly.
}
}
}
