% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_dynamic_threshold.R
\name{DynamicThresholdScorer}
\alias{DynamicThresholdScorer}
\title{Dynamic Threshold Scorer}
\description{
R6 class that implements the online anomaly scoring method (based on the prediction errors)
proposed by Buda et al.
}
\examples{

scorer <- DynamicThresholdScorer$new(3)

scorer$computeScore(0.5)
scorer$computeScore(0.6)
scorer$computeScore(0.5)
scorer$computeScore(7)
scorer$computeScore(0.2)

}
\references{
T. S. Buda, B. Caglayan, H. Assem,  DeepAD: A generic frameworkbased on deep
learning for time series anomaly detection, in: LectureNotes in Computer Science
(including subseries Lecture Notes in Artificial Intelligence and Lecture Notes in
Bioinformatics), volume 10937LNAI, Springer Verlag, 2018, pp. 577-588
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DynamicThresholdScorer$new()}}
\item \href{#method-computeScore}{\code{DynamicThresholdScorer$computeScore()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new DynamicThresholdScorer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicThresholdScorer$new(wnMin = 100, wnMax = 2000, l = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wnMin}}{Minimum window size to start computing the anomaly score. It must be an integer
between [1, \code{wnMax}].}

\item{\code{wnMax}}{Maximum window size to compute the anomaly score. It must be an integer
equal or bigger than \code{wnMin}.}

\item{\code{l}}{Times standard deviation. It must be an integer greater than 0. By default 10, but
other common values could be 3 and 6.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new DynamicThresholdScorer object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-computeScore"></a>}}
\if{latex}{\out{\hypertarget{method-computeScore}{}}}
\subsection{Method \code{computeScore()}}{
Calculates the anomaly score from the prediction error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicThresholdScorer$computeScore(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Current error value to be scored.}

\item{\code{...}}{Any other parameter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Anomaly score, a value between [0,1]. The closer the number is
to 1, the higher the chance it is an anomaly.
}
}
}
