% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ec2_cpu_utilization_53ea38}
\alias{ec2_cpu_utilization_53ea38}
\title{ec2_cpu_utilization_53ea38.}
\format{
A data frame with three variables: \code{timestamp}, \code{value},
  \code{is.real.anomaly}.

For further details, see \url{https://github.com/numenta/NAB/blob/master/data/README.md}
}
\usage{
ec2_cpu_utilization_53ea38
}
\description{
AWS server metrics as collected by the AmazonCloudwatch service. Example metrics include
CPU Utilization, Network Bytes In, and Disk Read Bytes..
}
\keyword{datasets}
