% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ec2_request_latency_system_failure}
\alias{ec2_request_latency_system_failure}
\title{ec2_request_latency_system_failure.}
\format{
A data frame with three variables: \code{timestamp}, \code{value},
  \code{is.real.anomaly}.

For further details, see \url{https://github.com/numenta/NAB/blob/master/data/README.md}
}
\usage{
ec2_request_latency_system_failure
}
\description{
CPU usage data from a server in Amazon's East Coast datacenter. The dataset ends
with complete system failure resulting from a documented failure of AWS API
servers.
}
\keyword{datasets}
