% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elb_request_count_8c0756}
\alias{elb_request_count_8c0756}
\title{elb_request_count_8c0756.}
\format{A data frame with three variables: \code{timestamp}, \code{value},
  \code{is.real.anomaly}.

For further details, see \url{https://github.com/numenta/NAB/blob/master/data/README.md}}
\usage{
elb_request_count_8c0756

elb_request_count_8c0756
}
\description{
AWS server metrics as collected by the AmazonCloudwatch service. Example metrics include
CPU Utilization, Network Bytes In, and Disk Read Bytes..

AWS server metrics as collected by the AmazonCloudwatch service. Example metrics include
CPU Utilization, Network Bytes In, and Disk Read Bytes..
}
\keyword{datasets}
