% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exchange_3_cpm_results}
\alias{exchange_3_cpm_results}
\title{exchange_3_cpm_results.}
\format{
A data frame with three variables: \code{timestamp}, \code{value},
  \code{is.real.anomaly}.

For further details, see \url{https://github.com/numenta/NAB/blob/master/data/README.md}
}
\usage{
exchange_3_cpm_results
}
\description{
Online advertisement clicking rates, where the metrics are cost-per-click (CPC) and cost
per thousand impressions (CPM). One of the files is normal, without anomalies.
}
\keyword{datasets}
