% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnomalyScore.CAD.R
\name{GetAnomalyScore.CAD}
\alias{GetAnomalyScore.CAD}
\title{Get Anomaly Score for Contextual Anomaly Detector}
\usage{
GetAnomalyScore.CAD(datum, local.environment)
}
\arguments{
\item{datum}{Integer vector of activated bits.}

\item{local.environment}{Local environment with the variables of
the Context Anomaly Detector.}
}
\value{
Anomaly score of \code{datum}.
}
\description{
\code{GetAnomalyScore.CAD} calculates the anomaly score of a
single datum using the notion of contexts conformed by facts and provides
probabilistic abnormality scores.
}
\details{
\code{datum} must be a numerical vector of the transformation made by \code{IntToBinarySens}.
}
\references{
Smirnov, M. (2018). CAD: Contextual Anomaly
    Detector. https://github.com/smirmik/CAD
}
