% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetContextByFacts.CAD.R
\name{GetContextByFacts.CAD}
\alias{GetContextByFacts.CAD}
\title{The function which determines by the complete facts list whether the context
is already saved to the memory. If the context is not found the function
immediately creates such. To optimize speed and volume of the occupied memory
the contexts are divided into semi-contexts as several contexts can contain
the same facts set in its left and right parts.}
\usage{
GetContextByFacts.CAD(context.list, context.operator, zero.level = 0)
}
\arguments{
\item{newContextsList}{list of potentially new contexts}

\item{zerolevel}{flag indicating the context type in transmitted list}
}
\value{
Depending on the type of  potentially new context transmitted as
an input parameters the function returns either:
  а) flag indicating that the transmitted zero-level context is
a new/existing one;
or:
  b) number of the really new contexts that have been saved to the
context memory.
}
\description{
The function which determines by the complete facts list whether the context
is already saved to the memory. If the context is not found the function
immediately creates such. To optimize speed and volume of the occupied memory
the contexts are divided into semi-contexts as several contexts can contain
the same facts set in its left and right parts.
}
