% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDetectorScore_GetLabels.R
\name{GetLabels}
\alias{GetLabels}
\title{Get Lables}
\usage{
GetLabels(data)
}
\arguments{
\item{data}{All dataset with training and test datasets with at least \code{timestamp},
\code{value}, \code{is.anomaly}, \code{is.real.anomaly}, \code{start.limit} and
\code{end.limit} columns.}
}
\value{
Same data set with two additional columns \code{label} and \code{first.tp}.
\code{first.tp} indicates for each window Which is the position of first true positive.
\code{label} indicates for each detection if it is a TP, FP, TN or FN.
}
\description{
\code{GetLabels} Calculates the start and end positions of each window that
are focused on the real anomalies. This windows can be used to know if the detected anomaly is a
true positive or not.
}
\details{
\code{data} must be a data.frame with  \code{timestamp}, \code{value}, \code{is.anomaly}
and \code{is.real.anomaly} columns. \code{timestamp} column can be numeric, of type POSIXct, or a
character type date convertible to POSIXct. see  \code{\link{GetWindowsLimits}} to know more
about how to get \code{start.limit} and \code{end.limit} columns.
}
\references{
A. Lavin and S. Ahmad, “Evaluating Real-time Anomaly Detection Algorithms – the
Numenta Anomaly Benchmark,” in 14th International Conference on Machine Learning and
Applications (IEEE ICMLA’15), 2015.
}
