% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iio_us_east1_i_a2eb1cd9_NetworkIn}
\alias{iio_us_east1_i_a2eb1cd9_NetworkIn}
\title{iio_us_east1_i_a2eb1cd9_NetworkIn.}
\format{A data frame with three variables: \code{timestamp}, \code{value},
  \code{is.real.anomaly}.

For further details, see \url{https://github.com/numenta/NAB/blob/master/data/README.md}}
\usage{
iio_us_east1_i_a2eb1cd9_NetworkIn
}
\description{
AWS server metrics as collected by the AmazonCloudwatch service. Example metrics include
CPU Utilization, Network Bytes In, and Disk Read Bytes..
}
\keyword{datasets}
