% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_modules.R
\name{install_bencode}
\alias{install_bencode}
\title{Install python module bencode}
\usage{
install_bencode(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Argument for reticulate::py_install. Installation method. By default, it finds a method for the local environment.}

\item{conda}{Argument for reticulate::py_install. Path to conda executable.}
}
\description{
\code{install_bencode} installs bencode module.
}
\details{
On Linux and OS X the "virtualenv" method will be used by default ("conda" will be used if virtualenv isn't available). On Windows, the "conda" method is always used.
}
