% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntToBinarySens.R
\name{IntToBinarySens}
\alias{IntToBinarySens}
\title{Int To Binary Sens}
\usage{
IntToBinarySens(x, num.norm.value.bits)
}
\arguments{
\item{x}{input data}

\item{num.norm.value.bits}{Expected number of binary bins}
}
\value{
Vector with activated numeric bits.
}
\description{
\code{IntToBinarySens} Auxiliar function that transforms
integer to binary bins.
}
\examples{
otsad:::IntToBinarySens(2, 3)
otsad:::IntToBinarySens(10, 5)
}
