% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_taxi}
\alias{nyc_taxi}
\title{nyc_taxi.}
\format{
A data frame with three variables: \code{timestamp}, \code{value},
  \code{is.real.anomaly}.

For further details, see \url{https://github.com/numenta/NAB/blob/master/data/README.md}
}
\usage{
nyc_taxi
}
\description{
Number of NYC taxi passengers, where the five anomalies occur during the NYC
marathon, Thanksgiving, Christmas, New Years day, and a snow storm. The raw
data is from the NYC Taxi and Limousine Commission. The data file included
here consists of aggregating the total number of taxi passengers into 30
minute buckets.
}
\keyword{datasets}
