% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rds_cpu_utilization_cc0c53}
\alias{rds_cpu_utilization_cc0c53}
\title{rds_cpu_utilization_cc0c53.}
\format{
A data frame with three variables: \code{timestamp}, \code{value},
  \code{is.real.anomaly}.

For further details, see \url{https://github.com/numenta/NAB/blob/master/data/README.md}
}
\usage{
rds_cpu_utilization_cc0c53
}
\description{
AWS server metrics as collected by the AmazonCloudwatch service. Example metrics include
CPU Utilization, Network Bytes In, and Disk Read Bytes..
}
\keyword{datasets}
