% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReduceAnomalies.R
\name{ReduceAnomalies}
\alias{ReduceAnomalies}
\title{Reduce Anomalies}
\usage{
ReduceAnomalies(data, windowLength)
}
\arguments{
\item{data}{Numerical vector with anomaly labels.}

\item{windowLength}{Window length.}
}
\value{
New Numerical vector with reduced anomaly labels.
}
\description{
\code{ReduceAnomalies} It reduces the number of detected anomalies. This function is
designed to reduce the number of false positives keeping only the first detection of all those
that are close to each other. This proximity distance is defined by a window
}
