% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rogue_agent_key_updown}
\alias{rogue_agent_key_updown}
\title{rogue_agent_key_updown.}
\format{A data frame with three variables: \code{timestamp}, \code{value},
  \code{is.real.anomaly}.

For further details, see \url{https://github.com/numenta/NAB/blob/master/data/README.md}}
\usage{
rogue_agent_key_updown
}
\description{
Timing the key strokes for several users of a computer, where the anomalies
represent a change in the user.
}
\keyword{datasets}
