% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_sigma_scorer.R
\name{SigmaScorer}
\alias{SigmaScorer}
\title{Sigma scorer}
\description{
R6 class that allows computing the anomaly score online based on historic prediction errors and
3-sigma control limits.
}
\examples{

scorer <- SigmaScorer$new(3, 4)

scorer$computeScore(0.5)
scorer$computeScore(0.6)
scorer$computeScore(0.5)
scorer$computeScore(1)
scorer$computeScore(0.2)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SigmaScorer$new()}}
\item \href{#method-computeScore}{\code{SigmaScorer$computeScore()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new SigmaScorer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SigmaScorer$new(wnMin = 100, wnMax = 2000, l = 3, computeLogLikelihood = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wnMin}}{Minimum window size to start computing the anomaly score. It must be an integer
between [1, \code{wnMax}].}

\item{\code{wnMax}}{Maximum window size to compute the anomaly score. It must be an integer
equal or bigger than \code{wnMin}.}

\item{\code{l}}{Number of deviations from the mean. It must be an integer greater than 0. By default
3, but another common value is 6.}

\item{\code{computeLogLikelihood}}{If \code{TRUE} compute a log scale representation of the
likelihood value. Since thelikelihood computations return low probabilities that often go
into four 9's or five 9's, a log value is more useful for visualization, thresholding, etc.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new SigmaScorer object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-computeScore"></a>}}
\if{latex}{\out{\hypertarget{method-computeScore}{}}}
\subsection{Method \code{computeScore()}}{
Calculates the anomaly score from the prediction error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SigmaScorer$computeScore(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Current error value to be scored.}

\item{\code{...}}{Any other parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Anomaly score, a value between [0,1]. The closer the number is
to 1, the higher the chance it is an anomaly.
}
}
}
