% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step.CAD.R
\name{Step.CAD}
\alias{Step.CAD}
\title{Step operator for Contextual Anomaly Detector (ContextualAD)}
\usage{
Step.CAD(datum, local.environment)
}
\arguments{
\item{datum}{New datum in form of numeric vector corresponding to
the bits that have been activated.}

\item{local.environment}{Local environment with the variables of
the Contextual Anomaly Detector.}
}
\value{
If \code{left.or.right}, number of new contex added tho the
    \code{context.operator}. Otherwise, the output from
    \code{UpdateContextsAndGetActive.CAD}.
}
\description{
\code{Step.CAD} is an auxiliar method that makes the necessary
changes in the local environment of the anomaly detector
}
\details{
\code{datum} must be a numeric vector.
}
\references{
Smirnov, M. (2018). CAD: Contextual Anomaly
    Detector. https://github.com/smirmik/CAD
}
