% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TravelTime_451}
\alias{TravelTime_451}
\title{TravelTime_451.}
\format{A data frame with three variables: \code{timestamp}, \code{value},
  \code{is.real.anomaly}.

For further details, see \url{https://github.com/numenta/NAB/blob/master/data/README.md}}
\usage{
TravelTime_451
}
\description{
Real time traffic data from the Twin Cities Metro area in Minnesota, collected
by the Minnesota Department of Transportation. Included metrics include
occupancy, speed, and travel time from specific sensors.
}
\keyword{datasets}
