% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Twitter_volume_AMZN}
\alias{Twitter_volume_AMZN}
\title{Twitter_volume_AMZN.}
\format{A data frame with three variables: \code{timestamp}, \code{value},
  \code{is.real.anomaly}.

For further details, see \url{https://github.com/numenta/NAB/blob/master/data/README.md}}
\usage{
Twitter_volume_AMZN
}
\description{
A collection of Twitter mentions of large publicly-traded companies such
as Google and IBM. The metric value represents the number of mentions for
a given ticker symbol every 5 minutes.
}
\keyword{datasets}
