% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateContextsAndGetActive.CAD.R
\name{UpdateContextsAndGetActive.CAD}
\alias{UpdateContextsAndGetActive.CAD}
\title{Update Contexts and Get Active - Contextual Anomaly Detector}
\usage{
UpdateContextsAndGetActive.CAD(new.ctxt.flag, context.operator)
}
\arguments{
\item{new.ctxt.flag}{flag indicating that a new zero-level context
is not recorded at the current step, which means that all
contexts already exist and there is no need to create new ones.}

\item{context.operator}{Environment with the current status for the
context operator.}
}
\value{
List of:

\item{activeContexts}{List of identifiers of the contexts which completely coincide with the input stream, should be considered active and be recorded to the input stream of “neurons”}
\item{potentialNewContextList}{List of contexts based on intersection between the left and the right zero-level semi-contexts, which are potentially new contexts requiring saving to the context memory}
}
\description{
\code{UpdateContextsAndGetActive.CAD} is an auxiliar method for the
Contextual Anomaly Detector that reviews the list of previously
selected left semi-contexts, creates the list of potentially new
contexts resulted from intersection between zero-level contexts,
determines the contexts that coincide with the input data and
require activation.
}
\references{
Smirnov, M. (2018). CAD: Contextual Anomaly
    Detector. https://github.com/smirmik/CAD
}
