% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer_window_normalizer.R
\name{WindowNormalizer}
\alias{WindowNormalizer}
\title{Window-based streaming normalizalizer}
\description{
R6 class allows normalizing the data set in an online manner one-by-one or chunk-by-chunk. It
maintains a window with the last data values to calculate the statistical values needed for the
min-max normalization or the z-score standardization.
}
\examples{

normalizer <- WindowNormalizer$new(wl = 3)

normalizer$normalize(10)
normalizer$normalize(15)
normalizer$normalize(20)
normalizer$normalize(10)
normalizer$normalize(30)
normalizer$normalize(15)

normalizer$denormalize(-0.3202563)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WindowNormalizer$new()}}
\item \href{#method-normalize}{\code{WindowNormalizer$normalize()}}
\item \href{#method-denormalize}{\code{WindowNormalizer$denormalize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new WindowNormalizer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WindowNormalizer$new(wl, method = "z-score", returnPoint = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wl}}{Length of the window with historical data values.}

\item{\code{method}}{Normalization method to be used "min-max" or "z-score".}

\item{\code{returnPoint}}{If \code{FALSE} then normalized current window is returned, else,
normalized current data point is returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new WindowNormalizer object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-normalize"></a>}}
\if{latex}{\out{\hypertarget{method-normalize}{}}}
\subsection{Method \code{normalize()}}{
Normalizes the current data value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WindowNormalizer$normalize(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Current data value to be normalized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If \code{returnPoint = FALSE} then normalized current window is returned, else,
normalized current data point is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-denormalize"></a>}}
\if{latex}{\out{\hypertarget{method-denormalize}{}}}
\subsection{Method \code{denormalize()}}{
Denormalizes the current data value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WindowNormalizer$denormalize(y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{Current data value to be denormalized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Denormalized current data point.
}
}
}
