% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDetectorScore_normalizeScore.R
\name{NormalizeScore}
\alias{NormalizeScore}
\title{Normalize Score using Max and Min normalization}
\usage{
NormalizeScore(real.score, perfect.score, null.score)
}
\arguments{
\item{real.score}{Detector score. See \code{\link{GetDetectorScore}}.}

\item{perfect.score}{Perfect detector score; one that outputs all true positives and no false
positives. See \code{\link{GetNullAndPerfectScores}}.}

\item{null.score}{Perfect detector score; one that outputs all true positives and no false
positives. See \code{\link{GetNullAndPerfectScores}}.}
}
\value{
Normalized score.
}
\description{
\code{ReduceAnomalies} It reduces the number of detected anomalies. This function is
designed to reduce the number of false positives keeping only the first detection of all those
that are close to each other. This proximity distance is defined by a window
}
\references{
A. Lavin and S. Ahmad, “Evaluating Real-time Anomaly Detection Algorithms – the
Numenta Anomaly Benchmark,” in 14th International Conference on Machine Learning and
Applications (IEEE ICMLA’15), 2015.
}
