% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocp_knn_cad.R
\name{OcpKnnCad}
\alias{OcpKnnCad}
\title{Optimized Clasical processing KNN based Conformal Anomaly Detector}
\usage{
OcpKnnCad(data, threshold, l, n, m, k)
}
\arguments{
\item{data}{all data set}

\item{threshold}{detection threshold}

\item{l}{window length}

\item{n}{number of training rows}

\item{m}{number of calibration rows}

\item{k}{number of neighbours to take into account}
}
\value{
anomaly and grade results
}
\description{
calculates the anomalies of a data set using the optimized version of the classical processing based on the KNN-CAD algorithm
}
