% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateContextsAndGetActive.CAD.R
\name{UpdateContextsAndGetActive.CAD}
\alias{UpdateContextsAndGetActive.CAD}
\title{This function reviews the list of previously selected left semi-contexts,
creates the list of potentially new contexts resulted from intersection
between zero-level contexts, determines the contexts that coincide with
the input data and require activation.}
\usage{
UpdateContextsAndGetActive.CAD(new.ctxt.flag, context.operator)
}
\arguments{
\item{newContextFlag:}{flag indicating that a new zero-level
context is not recorded at the current
step, which means that all contexts
already exist and there is no need to
create new ones.}
}
\value{
activeContexts:         list of identifiers of the contexts which
completely coincide with the input stream,
should be considered active and be 
recorded to the input stream of “neurons”

potentialNewContextList:  list of contexts based on intersection
between the left and the right zero-level
semi-contexts, which are potentially new
contexts requiring saving to the context
memory
}
\description{
This function reviews the list of previously selected left semi-contexts,
creates the list of potentially new contexts resulted from intersection
between zero-level contexts, determines the contexts that coincide with
the input data and require activation.
}
